+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- split Space keycode in action to SpaceLeft and SpaceRight
+
 ## [0.0.43](https://github.com/andy23512/tangent-cc-lib/releases/tag/0.0.43)
 
 - add no break space action